Plot an ATC Preview as a Tree Diagram

Due to the high amount of records, use preview_atc_classification to determine the appropriate range based on row counts per level to supply the optional range argument. The final output is either 1. the data.tree-style dataframe if the skip_plot argument is set to true, 2. a plot in the Viewer, or 3. a plot saved to an html file as named by the file argument if one is provided.

Due to the high amount of records, use preview_atc_classification to determine the appropriate range based on row counts per level to supply the optional range argument. The final output is either 1. the data.tree-style dataframe if the skip_plot argument is set to true, 2. a plot in the Viewer, or 3. a plot saved to an html file as named by the file argument if one is provided.

Plot a LOINC Class. Due to the high amount of records, use loinc_classification to determine the appropriate range based on row counts per level to supply the optional range argument.

plot_atc_classification(
  conn,
  concept_obj,
  range,
  file,
  vocab_schema = "omop_vocabulary",
  color_by = "standard_concept",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)

plot_atc_classification(
  conn,
  concept_obj,
  range,
  file,
  vocab_schema = "omop_vocabulary",
  color_by = "standard_concept",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)

plot_atc_classification(
  conn,
  concept_obj,
  range,
  file,
  vocab_schema = "omop_vocabulary",
  color_by = "standard_concept",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)

Arguments

skip_plot
If true, returns the dataframe before it is plotted and plotting is not done. This is an option to troubleshoot or customize a plot beyond what is available within the function.

See also